fix: pull 에러 수정
