+++ src/components/Router.tsx
@@ -4,7 +4,7 @@ import { AddBook, Bookcase, Login, Main, MyPage, Signup } from "../pages";
 import BookNote from "../pages/BookNote";
 import ToBe from "../pages/ToBe";
 import { PeriRead, PostRead, PreRead, Total } from "./bookcase";
-import PeriNote from "./bookNote/periNote/PeriNote";
+import { PeriNote } from "./bookNote";
 import PreNote from "./bookNote/preNote/PreNote";
 import { CommonLayout } from "./common";
 import { FirstStep, LastStep, SecondStep, ThirdStep } from "./signup";
