+++ src/components/bookNote/index.ts
@@ -5,7 +5,7 @@ export { default as DrawerWrapper } from "./common/DrawerWrapper";
 export { default as ExButton } from "./common/ExButton";
 export { default as StepUp } from "./common/StepUp";
 export { default as Navigator } from "./Navigator";
-export { default as PeriNote } from "./periNote/PeriNote";
+export { default as PeriNote } from "./perinote/PeriNote";
 export { default as InputQuestion } from "./preNote/InputQuestion";
 export { default as PreNoteForm } from "./preNote/PreNoteForm";
 export { default as QuestionThree } from "./preNote/QuestionThree";
